Extract cranelift intrinsics and enforce the monolith ratchet in CI
First real cranelift_backend.rs split under #1254: the pure
runtime-intrinsic implementations (JSON scalar parse/stringify, the
stage1-safe regex engine and its NFA types, percent encoding, and the
crypto primitives with their opaque FFI handle types) move to
stage1/crates/axiomc/src/cranelift_backend/intrinsics.rs. The module is
independent of evaluator and lowering state; call sites are unchanged
via a crate-visible re-export.

Wire the ratchet into the fast PR lane: run-fast-checks.sh now runs
report-compiler-source-monoliths.py --check-plan --check-ratchet
against the PR data checkout (--source-root/--plan follow the #1360
checkout-root pattern) plus its self-test, so monolith growth fails CI
unless the ceiling change is explicit in the same PR. The plan records
the lowered cranelift ceiling, the new intrinsics ceiling, and explicit
raises for the tracked files that grew while the ratchet was still
advisory (merged #1355-#1376 era), with the rationale in the narrative.

Refs #1254.
