+++ scripts/ci/run-fast-checks.sh
@@ -34,6 +34,18 @@ if ! python3 "$script_repo_root/scripts/ci/check-direct-native-runtime-abi.py" \
   exit 1
 fi
 rm -f "$abi_matrix_report"
+monolith_report="$(mktemp)"
+if ! python3 "$script_repo_root/scripts/ci/report-compiler-source-monoliths.py" \
+  --source-root "$repo_root/stage1/crates/axiomc/src" \
+  --plan "$repo_root/docs/compiler-source-decomposition-plan.md" \
+  --json --check-plan --check-ratchet >"$monolith_report"; then
+  echo "compiler source monolith ratchet failed (#1254); shrink the file or update the ceiling in docs/compiler-source-decomposition-plan.md in this PR:" >&2
+  python3 -c 'import json,sys; d=json.load(open(sys.argv[1])); [print(e, file=sys.stderr) for e in d.get("plan_check",{}).get("errors",[])+d.get("ratchet_check",{}).get("errors",[])]' "$monolith_report" || cat "$monolith_report" >&2
+  rm -f "$monolith_report"
+  exit 1
+fi
+rm -f "$monolith_report"
+python3 "$script_repo_root/scripts/ci/test-report-compiler-source-monoliths.py"
 bash "$script_repo_root/scripts/ci/test-check-rust-exit-command-surface.sh"
 python3 "$script_repo_root/scripts/ci/test-pr-queue-remediation.py"
 python3 "$script_repo_root/scripts/ci/test-report-delivery-signals.py"
